# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthRequest {
  /** ID do cliente fornecido pela Teck Sign */
  clientId: string;
  /** Secret do cliente fornecido pela Teck Sign */
  clientSecret: string;
}

export interface AuthResponse {
  /** Token JWT de acesso */
  accessToken: string;
  /** Tipo do token */
  tokenType: string;
  /** Validade do token em segundos */
  expiresIn: number;
}

export interface Signatory {
  /** Nome completo do signatário */
  name: string;
  /** E-mail do signatário */
  email: string;
  /** CPF do signatário (apenas dígitos) */
  cpf?: string;
  /** Ordem de assinatura (1 = primeiro) */
  signingOrder?: number;
}

export interface CreateProcessRequest {
  /** Título do processo */
  title: string;
  /** Mensagem enviada aos signatários por e-mail */
  message?: string;
  /** Documento PDF codificado em Base64 */
  documentBase64: string;
  /** Lista de signatários do processo */
  signatories: Signatory[];
  /** Data de expiração do processo */
  expiresAt?: string;
}

/**
 * Status do processo
 */
export type CreateProcessResponseStatus =
  (typeof CreateProcessResponseStatus)[keyof typeof CreateProcessResponseStatus];

export const CreateProcessResponseStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
  expired: "expired",
} as const;

export interface CreateProcessResponse {
  /** ID único do processo criado */
  processId: string;
  /** Status do processo */
  status: CreateProcessResponseStatus;
  /** Data de criação do processo */
  createdAt: string;
  /** URL para acompanhar o processo */
  signingUrl?: string;
}

export type SignatoryStatusStatus =
  (typeof SignatoryStatusStatus)[keyof typeof SignatoryStatusStatus];

export const SignatoryStatusStatus = {
  pending: "pending",
  viewed: "viewed",
  signed: "signed",
  refused: "refused",
} as const;

export interface SignatoryStatus {
  name: string;
  email: string;
  status: SignatoryStatusStatus;
  signedAt?: string;
}

export type ProcessResponseStatus =
  (typeof ProcessResponseStatus)[keyof typeof ProcessResponseStatus];

export const ProcessResponseStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
  expired: "expired",
} as const;

export interface ProcessResponse {
  processId: string;
  title: string;
  status: ProcessResponseStatus;
  createdAt: string;
  expiresAt?: string;
  signatories: SignatoryStatus[];
  totalSignatories: number;
  signedCount: number;
}

/**
 * original = documento enviado; signed = documento com assinaturas
 */
export type ProcessDocumentType =
  (typeof ProcessDocumentType)[keyof typeof ProcessDocumentType];

export const ProcessDocumentType = {
  original: "original",
  signed: "signed",
} as const;

export interface ProcessDocument {
  documentId: string;
  name: string;
  /** original = documento enviado; signed = documento com assinaturas */
  type: ProcessDocumentType;
  /** URL temporária para download (válida por 60 minutos) */
  downloadUrl: string;
  sizeBytes?: number;
  createdAt: string;
}

export interface ProcessDocumentsResponse {
  processId: string;
  documents: ProcessDocument[];
}

export interface ErrorResponse {
  statusCode: number;
  error: string;
  message: string;
}