# lib/api-zod/src/generated/types/authRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */

export interface AuthRequest {
  /** ID do cliente fornecido pela Teck Sign */
  clientId: string;
  /** Secret do cliente fornecido pela Teck Sign */
  clientSecret: string;
}