# lib/api-zod/src/generated/types/createProcessRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */
import type { Signatory } from "./signatory";

export interface CreateProcessRequest {
  /** Título do processo */
  title: string;
  /** Mensagem enviada aos signatários por e-mail */
  message?: string;
  /** Documento PDF codificado em Base64 */
  documentBase64: string;
  /** Lista de signatários do processo */
  signatories: Signatory[];
  /** Data de expiração do processo */
  expiresAt?: Date;
}