# lib/api-zod/src/generated/types/createProcessResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */
import type { CreateProcessResponseStatus } from "./createProcessResponseStatus";

export interface CreateProcessResponse {
  /** ID único do processo criado */
  processId: string;
  /** Status do processo */
  status: CreateProcessResponseStatus;
  /** Data de criação do processo */
  createdAt: Date;
  /** URL para acompanhar o processo */
  signingUrl?: string;
}