# lib/api-zod/src/generated/types/processDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */
import type { ProcessDocumentType } from "./processDocumentType";

export interface ProcessDocument {
  documentId: string;
  name: string;
  /** original = documento enviado; signed = documento com assinaturas */
  type: ProcessDocumentType;
  /** URL temporária para download (válida por 60 minutos) */
  downloadUrl: string;
  sizeBytes?: number;
  createdAt: Date;
}