# lib/api-zod/src/generated/types/processDocumentsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */
import type { ProcessDocument } from "./processDocument";

export interface ProcessDocumentsResponse {
  processId: string;
  documents: ProcessDocument[];
}