# lib/api-zod/src/generated/types/processResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */
import type { ProcessResponseStatus } from "./processResponseStatus";
import type { SignatoryStatus } from "./signatoryStatus";

export interface ProcessResponse {
  processId: string;
  title: string;
  status: ProcessResponseStatus;
  createdAt: Date;
  expiresAt?: Date;
  signatories: SignatoryStatus[];
  totalSignatories: number;
  signedCount: number;
}