# lib/api-zod/src/generated/types/signatory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */

export interface Signatory {
  /** Nome completo do signatário */
  name: string;
  /** E-mail do signatário */
  email: string;
  /** CPF do signatário (apenas dígitos) */
  cpf?: string;
  /** Ordem de assinatura (1 = primeiro) */
  signingOrder?: number;
}