# lib/api-zod/src/generated/types/signatoryStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teck Sign API - Plataforma de Assinatura Digital
 * OpenAPI spec version: 1.0.0
 */
import type { SignatoryStatusStatus } from "./signatoryStatusStatus";

export interface SignatoryStatus {
  name: string;
  email: string;
  status: SignatoryStatusStatus;
  signedAt?: Date;
}